- fix error on register dio in getIt - extract method from dio factory - add update Quantity(cart) with retrofit
