+++ lib/core/api/new_api_manger.dart
@@ -1,13 +1,14 @@
 import 'package:dartz/dartz.dart';
 import 'package:dio/dio.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
-import 'package:e_comarce_clean/features/auth/data/models/UserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/response/UserModel.dart';
 import 'package:e_comarce_clean/features/category_feature/data/models/SubCategoryModel.dart';
 import 'package:e_comarce_clean/features/home/data/models/BrandModel.dart';
 import 'package:injectable/injectable.dart';
 import 'package:retrofit/error_logger.dart';
 import 'package:retrofit/http.dart';
 
+import '../../features/auth/data/models/request/SignUserModel.dart';
 import '../../features/cart_feature/data/models/CartProductsModel.dart';
 import '../../features/home/data/models/CategoryModel.dart';
 import '../../features/products_feature/data/models/ProductModel.dart';
@@ -28,13 +29,13 @@ abstract class NewApiManger {
 
   @GET(EndPoints.getAllBrand)
   Future<BrandModel> getAllBrand();
-
+/////////////category/////////////////////
   @GET(EndPoints.getProductBaseOnCategory)
   Future<ProductModel> getProductBaseOnCategory(@Path() String categoryId);
   @GET(EndPoints.getProductBaseOnCategory)
   Future<SubCategoryModel> getProductBaseOnCategoryForCategory(@Path() String categoryId);
 
-
+///////////////cart/////////////////////
   @GET(EndPoints.getCartProduct)
   Future<CartProductsModel> getCartProduct();
 
@@ -43,13 +44,16 @@ abstract class NewApiManger {
 
   @DELETE(EndPoints.deleteFromCart)
   Future<void> deleteFromCart(@Path() String productId);
+  ///////////wishlist/////////////////////
   @GET(EndPoints.getWishlistData)
   Future<WishlistProductModel> getWishlistData();
   @POST(EndPoints.addToWishList)
   Future<void> addToWishList(@Body() Map<String, dynamic> body);
   @DELETE(EndPoints.deleteFromWishList)
   Future<void> deleteFromWishList(@Path() String productId);
-
+/////////////// auth/////////////////////
   @POST(EndPoints.logIn)
   Future<UserModel> signIn(@Body() Map<String, dynamic> body);
+  @POST(EndPoints.signUp)
+  Future<UserModel> signUp(@Body()SignUserModel signUserModel);
 }
