+++ lib/core/utils/helper_function.dart
@@ -1,3 +1,5 @@
+import 'package:dio/dio.dart';
+
 class HelperFunction {
   static String checkOnDiscount(num? price, num? priceAfterDiscount) {
     return priceAfterDiscount == null
@@ -11,6 +13,11 @@ class HelperFunction {
     } else {
       return true;
     }
+  }
+  static addTokenAfterLogin(String token,Dio dio) {
+    dio.options.headers = {
+      "token":token,
+    };
   }
