+++ lib/features/auth/data/data_sources/remote_data_source/remote_data_source.dart
@@ -1,8 +1,8 @@
 import 'package:dartz/dartz.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
-import 'package:e_comarce_clean/features/auth/data/models/SignUserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/request/SignUserModel.dart';
 
-import '../../models/UserModel.dart';
+import '../../models/response/UserModel.dart';
 
 abstract class AuthRemoteDataSource{
  Future<Either<Failure,UserModel>> signUp(SignUserModel signUserModel);
