+++ lib/features/auth/data/data_sources/remote_data_source/remote_datasource_impl.dart
@@ -5,9 +5,11 @@ import 'package:e_comarce_clean/core/api/dio_factory.dart';
 import 'package:e_comarce_clean/core/api/end_points.dart';
 import 'package:e_comarce_clean/core/api/new_api_manger.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
+import 'package:e_comarce_clean/core/service_locator/config.dart';
+import 'package:e_comarce_clean/core/utils/helper_function.dart';
 import 'package:e_comarce_clean/features/auth/data/data_sources/remote_data_source/remote_data_source.dart';
-import 'package:e_comarce_clean/features/auth/data/models/SignUserModel.dart';
-import 'package:e_comarce_clean/features/auth/data/models/UserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/request/SignUserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/response/UserModel.dart';
 import 'package:injectable/injectable.dart';
 
 import '../../../../../core/cache/storage_token.dart';
@@ -43,7 +45,7 @@ class AuthRemoteDataSourceImpl implements AuthRemoteDataSource {
 
       UserModel userModel = await newApiManger.signIn({"email": email, "password": password});
       storageToken.setToken(userModel.token??'');
-      DioFactory.addTokenAfterLogin(userModel.token??'');
+      HelperFunction.addTokenAfterLogin(userModel.token??"", getIt<Dio>());
       return right(userModel);
     } catch (e) {
       if (e is DioException) {
@@ -53,4 +55,5 @@ class AuthRemoteDataSourceImpl implements AuthRemoteDataSource {
       }
     }
   }
+
 }
