+++ lib/features/auth/data/repositories/auth_repo_impl.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
 import 'package:e_comarce_clean/features/auth/data/data_sources/remote_data_source/remote_data_source.dart';
-import 'package:e_comarce_clean/features/auth/data/models/SignUserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/request/SignUserModel.dart';
 
 import 'package:e_comarce_clean/features/auth/domain/repositories/auth_rebo.dart';
 import 'package:injectable/injectable.dart';
