+++ lib/features/auth/domain/repositories/auth_rebo.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
-import 'package:e_comarce_clean/features/auth/data/models/SignUserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/request/SignUserModel.dart';
 
 
 import '../entities/UserEntity.dart';
