+++ lib/features/auth/domain/use_cases/sign_up_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:e_comarce_clean/core/erorr/failure.dart';
-import 'package:e_comarce_clean/features/auth/data/models/SignUserModel.dart';
+import 'package:e_comarce_clean/features/auth/data/models/request/SignUserModel.dart';
 
 import 'package:e_comarce_clean/features/auth/domain/repositories/auth_rebo.dart';
 import 'package:injectable/injectable.dart';
